+++ Clients/VidPlayClient.py
@@ -75,10 +75,7 @@ class VidPlayClient(BaseClient):
         return decoded
 
     # step-4.4.1
-    def _encode_data(self, v_id: str) -> str:
-        resp = self._send_request(self.keys_url, return_type='json')['payload']['blob']['rawLines'][0]
-
-        key1, key2 = json.loads(resp)
+    def _encode_data(self, v_id: str, key1: str, key2: str) -> str:
         decoded_id = self._decode_data(key1, v_id)
         encoded_result = self._decode_data(key2, decoded_id)
 
@@ -125,17 +122,30 @@ class VidPlayClient(BaseClient):
         self.logger.debug(f'Available subtitles: {subtitles}')
         return subtitles
  
-    # step-4.4
-    def _resolve_sources(self, url: str):
+    # step-4.4.1
+    def _resolve_sources_inner(self, url: str, commit_id: str = 'main'):
         '''
-        resolve the download sources and return the list of download sources
+        resolve the download sources and return the list of download sources using given repo commit id.
         '''
-        self.logger.debug(f'Resolving sources from url: {url}')
+        self.logger.debug(f'Resolving sources for url [{url}] using commit id: {commit_id}')
         url_data = url.split("?")
 
         try:
+            # Extract keys from External Git Repo
+            keys_url = self.keys_url.format(commit_id=commit_id)
+            self.logger.debug(f'Fetching keys from external git source [Commit: {commit_id}], url: {keys_url}')
+            resp = self._send_request(keys_url, return_type='json')['payload']['blob']['rawLines'][0]
+            self.logger.debug(f'Fetched keys from {keys_url} are: {resp}')
+            key1, key2 = json.loads(resp)
+        except Exception as e:
+            self.logger.warning(f'No keys found for commit id: {commit_id}')
+            return {'error': 'vidsrc keys not found'}
+
+        try:
             # encode key required for futoken
-            key = self._encode_data(url_data[0].split("/e/")[-1])
+            self.logger.debug('Extracting the encoding key required for futoken...')
+            key = self._encode_data(url_data[0].split("/e/")[-1], key1, key2)
+            self.logger.debug(f'Extracted encoding key for futoken: {key}')
         except Exception as e:
             return {'error': f'Failed to encode key. Error: {e}'}
 
@@ -151,9 +161,54 @@ class VidPlayClient(BaseClient):
         req_data = self._send_request(vidplay_decoded_link, referer=url, return_type='json')
         self.logger.debug(f'Download sources response: {req_data}')
 
-        if req_data and req_data.get("result") and type(req_data.get("result")) == dict:
-            sources = req_data.get("result").get("sources")
-            self.logger.debug(f'Download sources: {sources}')
-            return sources
+        if req_data and req_data.get("result"):
+            if type(req_data.get("result")) == dict:
+                sources = req_data.get("result").get("sources")
+                self.logger.debug(f'Download sources: {sources}')
+                return sources
+            elif req_data.get("result") == 404:
+                self.logger.warning('Failed to fetch download sources with response code: 404. Probably keys expired!')
+                return {'error': 'Invalid vidsrc keys'}
 
         return {'error': 'No download links found'}
+
+    # step-4.4
+    def _resolve_sources(self, url: str):
+        '''
+        resolve the download sources and return the list of download sources.
+        '''
+        self.logger.debug(f'Start resolving sources using recursive approach')
+
+        # Resolving resources. Use main branch first
+        response = self._resolve_sources_inner(url, commit_id='main')
+        if not ('error' in response and 'vidsrc keys' in response.get('error')):
+            return response
+
+        self.logger.debug('Looking for vidsrc keys in alternate commits')
+
+        # idea is to look for vidsrc keys in the pull requests. Becoz even if owner is inactive, community is active :)
+        keys_base_url = '/'.join(self.keys_url.split('/')[:-3])
+        pull_soup = self._get_bsoup(f'{keys_base_url}/pulls', extra_headers={'Accept': 'text/html, application/xhtml+xml'}, silent=True)
+
+        available_prs = [ i['href'].split('/')[-1] for i in pull_soup.select('div a') if i.get('data-hovercard-type') == 'pull_request' ]
+        self.logger.debug(f'Available Pull Requests for Vidsrc keys: {available_prs}')
+
+        for pr in available_prs:
+            pr_url = f'{keys_base_url}/pull/{pr}/files'
+            try:
+                self.logger.debug(f'Extracting commit id from: {pr_url}')
+                soup = self._get_bsoup(f'{pr_url}', silent=True)
+                commit_id = [ i.get('data-commit') for i in soup.select('a.select-menu-item') if i.get('data-commit') ][0]
+                self.logger.debug(f'Extracted commit id: {commit_id}')
+                self.logger.info(f'Retrying to resolve sources with commit id: {commit_id}')
+                # Resolving resources. Use other alternatives.
+                response = self._resolve_sources_inner(url, commit_id=commit_id)
+                if not ('error' in response and 'vidsrc keys' in response.get('error')):
+                    return response
+            except Exception as e:
+                self.logger.debug(f'Failed to fetch commit id with error: {e}')
+
+        else:
+            self.logger.info('No other sources found for vidsrc keys!')
+
+        return response